extract StoreManager
